[Adjustment][Ezra] fix cycle require
